move unsupported device check to sdk, change theme name
