+++ internal-demo/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.PhysiologySdk">
+        android:theme="@style/Theme.SmartSpectra">
         <activity
             android:name=".MainActivity"
             android:exported="true">
