+++ internal-demo/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.PhysiologySdk" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.SmartSpectra" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.PhysiologySdk" parent="Base.Theme.PhysiologySdk" />
+    <style name="Theme.SmartSpectra" parent="Base.Theme.SmartSpectra" />
 </resources>
